requirements.txt
# requirements.txt
# Core ML
torch>=1.9.0
torchvision>=0.10.0
tensorboard>=2.7.0

# API & Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=0.5.0

# Utilities
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0

# Development
pytest>=6.2.0
black>=21.8b0
flake8>=3.9.0

#datasets
kaggle>=1.7.4.0